Opprydning i filstruktur
