--- frontend/arrangor-flate/app/components/Accordion.module.css
@@ -1,4 +1,0 @@
-/** Target <Accordion.Header> last span, to take up remaining space */
-.fullHeader span:last-child {
-    width: 100%;
-}
