+++ frontend/arrangor-flate/app/components/common/PageHeading.tsx
@@ -9,7 +9,7 @@ interface Props {
   };
 }
 
-export function PageHeader({ title, tilbakeLenke }: Props) {
+export function PageHeading({ title, tilbakeLenke }: Props) {
   return (
     <div className="flex flex-col relative gap-1">
       <Heading level="2" size="large" className="mb-3" id="innsending-table-header">
