+++ frontend/arrangor-flate/app/components/header/Header.tsx
@@ -1,8 +1,8 @@
 import { Link } from "react-router";
-import { Arrangorvelger } from "./arrangorvelger/Arrangorvelger";
 import { Arrangor } from "api-client";
 import { Heading, HStack } from "@navikt/ds-react";
 import { HeaderIcon } from "./HeaderIcon";
+import { Arrangorvelger } from "../arrangorvelger/Arrangorvelger";
 
 interface Props {
   arrangorer: Arrangor[];
