+++ frontend/arrangor-flate/app/components/header/HeaderIcon.tsx
