+++ frontend/arrangor-flate/app/components/utbetaling/UtbetalingTilsagnDetaljer.tsx
@@ -1,4 +1,4 @@
-import { Definisjonsliste } from "../Definisjonsliste";
+import { Definisjonsliste } from "../common/Definisjonsliste";
 import { ArrangorflateTilsagn } from "api-client";
 import { formaterNOK } from "@mr/frontend-common/utils/utils";
 import { formaterTilsagnType } from "../tilsagn/TilsagnTable";
