+++ frontend/arrangor-flate/app/routes/$orgnr._manuell-innsending.opprett-krav.oppsummering.tsx
@@ -22,8 +22,8 @@ import { destroySession, getSession } from "~/sessions.server";
 import { apiHeaders } from "~/auth/auth.server";
 import { formaterNOK, jsonPointerToFieldPath } from "@mr/frontend-common/utils/utils";
 import { useEffect, useRef } from "react";
-import { Definisjonsliste } from "~/components/Definisjonsliste";
-import { Separator } from "~/components/Separator";
+import { Definisjonsliste } from "~/components/common/Definisjonsliste";
+import { Separator } from "~/components/common/Separator";
 import { tekster } from "~/tekster";
 import { FileUpload, FileUploadHandler, parseFormData } from "@mjackson/form-data-parser";
 import { FileUploader } from "~/components/fileUploader/FileUploader";
