+++ frontend/arrangor-flate/app/utils/sort-by.ts
@@ -1,13 +1,6 @@
-import { useState } from "react";
-
 export type SortOrder = "ascending" | "descending";
 
 export type SortBySelector<T> = (a: T) => string | number | boolean | undefined | null;
-
-export interface SortState<TKey> {
-  orderBy: TKey;
-  direction: SortOrder;
-}
 
 export function sortBy<T extends object>(
   array: T[],
@@ -30,22 +23,4 @@ function comparator<T>(a: T, b: T): number {
     return 1;
   }
   return 0;
-}
-
-export function useSortState<TKey extends string>() {
-  const [sort, setSort] = useState<SortState<TKey> | undefined>();
-
-  const handleSort = (clickedKey: TKey) => {
-    if (sort?.orderBy === clickedKey) {
-      if (sort.direction === "ascending") {
-        setSort({ orderBy: clickedKey, direction: "descending" });
-      } else {
-        setSort(undefined);
-      }
-    } else {
-      setSort({ orderBy: clickedKey, direction: "ascending" });
-    }
-  };
-
-  return { sort, setSort, handleSort };
 }
